feat: Rename plural mods to singular in tx
